% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-inventory-shipment.R
\encoding{UTF-8}
\name{get_inventory_shipment_data}
\alias{get_inventory_shipment_data}
\title{get warehouse shipment data from BI database}
\usage{
get_inventory_shipment_data(con, brand_name, start_date, end_date, ...)
}
\arguments{
\item{con}{BI con connector}

\item{brand_name}{事业部}

\item{start_date}{时间周期}

\item{end_date}{时间周期}

\item{...}{出货量汇总字段}
}
\value{

}
\description{
获取仓库发货数据
}
\details{
出货数据包含 "加盟期货发货单", "加盟现货发货单", "直营调拨发货单","直发订单发货单",'加盟政策性发货单'五种单据，
YH_NO 要货单号,WBSTK 发货状态
当需要添加字段时可用BILL_TYPE 字段，其余门店属性商品属性自行添加
}
\examples{
get_inventory_shipment_data(con = con,brand_name = 'mujosh',
start_date = '2020-10-10',end_date = '2020-10-31',BILL_TYPE)

}
