% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ontheway-data.R
\encoding{UTF-8}
\name{get_ontheway_data}
\alias{get_ontheway_data}
\title{get on the way data from BI database}
\usage{
get_ontheway_data(
  con,
  brand_name,
  channel_type = NULL,
  area_name = NULL,
  boss_name = NULL,
  shop_no = NULL,
  category_name = NULL
)
}
\arguments{
\item{con}{BI connector}

\item{brand_name}{事业部名称}

\item{channel_type}{门店渠道}

\item{area_name}{管辖区域}

\item{boss_name}{加盟商客户}

\item{shop_no}{the shop of Number}

\item{category_name}{the category name of goods}
}
\value{
a data.frame
}
\description{
get on the way data from BI database
}
\details{
本在途数据仅仅是门店在途数据即总仓发货给门店的在途或门店发给总仓的在途两种类型，并不包含采购在途
}
\examples{
get_ontheway_data(con = con,brand_name = 'mujosh',category = c("frame","sunglasses"))

}
