% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-purchase-data.R
\encoding{UTF-8}
\name{get_purchase_data}
\alias{get_purchase_data}
\title{get purchase stock data from BI database}
\usage{
get_purchase_data(
  con,
  ...,
  start_date,
  end_date,
  brand_name,
  channel_type = NULL,
  area_name = NULL,
  boss_name = NULL,
  shop_no = NULL,
  category_name = NULL
)
}
\arguments{
\item{con}{BI connector}

\item{...}{汇总字段 可以按照 SHOP_NO,SKU_NO,年,月等}

\item{start_date}{数据开始日期}

\item{end_date}{数据结束日期}

\item{brand_name}{事业部名称}

\item{channel_type}{门店渠道}

\item{area_name}{管辖区域}

\item{boss_name}{加盟商客户}

\item{shop_no}{the shop of Number}

\item{category_name}{the category name of goods}
}
\value{

}
\description{
采购入库数据
}
\examples{
get_inventory_shipment_data(con = con,SKU_NO,brand_name = c('mujosh','aj'),
start_date = '2020-10-10',end_date = '2020-10-31',category = c("frame","sunglasses"))
get_purchase_data(con,SHOP_NO,brand_name = 'brand',start_date = '2021-01-01',end_date = '2021-01-10',category_name = 'frame')

}
