% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-shipment-data.R
\encoding{UTF-8}
\name{get_shipment_data}
\alias{get_shipment_data}
\title{Title get shipment data from Finance shipment table}
\usage{
get_shipment_data(con, brand_name, start_date, end_date, ...)
}
\arguments{
\item{con}{BI connector}

\item{brand_name}{事业部}

\item{start_date}{时间周期}

\item{end_date}{时间周期}

\item{...}{汇总字段 可以按照 SHOP_NO,SKU_NO,年,月等}
}
\value{
shipment data
}
\description{
从财务出货表中获取出货数据
}
\examples{
get_shipment_data(con,brand_name = '鏈ㄤ節鍗佷簨涓氶儴',start_date = '2020-01-01',end_date = '2020-10-25',骞<b4>,鏈<88>)


}
