% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-inventory.R
\encoding{UTF-8}
\name{get_store_inventory_data}
\alias{get_store_inventory_data}
\title{get store inventory from BI database}
\usage{
get_store_inventory_data(con, brand_name, stock_date, goods_categories, ...)
}
\arguments{
\item{con}{BI con connector}

\item{brand_name}{事业部}

\item{stock_date}{库存日期 默认昨天}

\item{goods_categories}{分析大类}

\item{...}{需要添加汇总字段,默认年,月,SHOP_NO,SKU_NO}
}
\value{
a data frame
}
\description{
get store inventory from BI database
}
\examples{
dt <- get_store_inventory_data(con = con,brand_name = 'mujosh',stock_date = lubridate::today()-days(1),goods_categories = c('lens'),SHOP_NO)
}
