% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-sales-stock-data.R
\encoding{UTF-8}
\name{merge_sales_stock_data}
\alias{merge_sales_stock_data}
\title{merge sales data  and stock data from BI database pro}
\usage{
merge_sales_stock_data(
  dates,
  con,
  brand_name,
  channel_type = NULL,
  area_name = NULL,
  boss_name = NULL,
  category_name = NULL,
  shop_no = NULL
)
}
\arguments{
\item{dates}{the list of date}

\item{con}{BI con connector}

\item{brand_name}{the name of brand}

\item{channel_type}{direct store and franchise store ,the channel of store sales}

\item{area_name}{area name}

\item{boss_name}{the franchise customer}

\item{category_name}{the category name of goods}

\item{shop_no}{the shop of Number}

\item{date}{the start date of you select}
}
\value{
a dataframe
}
\description{
full join sales data and stock data pro
}
\examples{
mydt <- merge_sales_stock_data(dates = c('2020-01-01','2020-10-11'),con,brand_name = "mujosh")
}
