% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sku.R
\encoding{UTF-8}
\name{sku}
\alias{sku}
\title{sku}
\usage{
sku(con, category_name = NULL)
}
\arguments{
\item{con}{BI con connector}

\item{category_name}{the category name of goods}
}
\value{
a dataframe
}
\description{
get the sku information from BI database
}
\examples{
sku <- sku(con)
}
