% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\encoding{UTF-8}
\name{store}
\alias{store}
\title{ghzy}
\usage{
store(
  con,
  brand_name,
  channel_type = NULL,
  area_name = NULL,
  boss_name = NULL,
  shop_no = NULL
)
}
\arguments{
\item{con}{BI con connector}

\item{brand_name}{brand_name}

\item{channel_type}{store type}

\item{area_name}{the area of store}

\item{shop_no}{the shop of Number}

\item{boo_name}{the boss of store}
}
\value{
a dataframe
}
\description{
get the store information from BI database
}
\examples{
store <- store(con,brand_name = "mujosh")
}
