% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-inventory-shipment.R
\encoding{UTF-8}
\name{get_inventory_shipment_data}
\alias{get_inventory_shipment_data}
\title{get warehouse shipment data from BI database}
\usage{
get_inventory_shipment_data(
  con,
  ...,
  start_date,
  end_date,
  brand_name,
  channel_type = NULL,
  area_name = NULL,
  boss_name = NULL,
  shop_no = NULL,
  category_name = NULL,
  WBSTK = "C"
)
}
\arguments{
\item{con}{BI connector}

\item{...}{汇总字段 可以按照 SHOP_NO,SKU_NO,年,月等}

\item{start_date}{数据开始日期}

\item{end_date}{数据结束日期}

\item{brand_name}{事业部名称}

\item{channel_type}{门店渠道}

\item{area_name}{管辖区域}

\item{boss_name}{加盟商客户}

\item{shop_no}{the shop of Number}

\item{category_name}{the category name of goods}

\item{WBSTK}{发货状态 默认为已经处理完成}
}
\value{
total warehouse shipment data.frame
}
\description{
get warehouse shipment data from BI database
}
\details{
出货数据包含 "加盟期货发货单", "加盟现货发货单", "直营调拨发货单","直发订单发货单",'加盟政策性发货单','加盟退货收货单'
'加盟免费发货单'等单据

YH_NO 要货单号,WBSTK 发货状态,默认为已经处理完成
当需要添加字段时可用BILL_TYPE 字段，其余门店属性商品属性自行添加
}
\examples{
get_inventory_shipment_data(
  con = con, BILL_TYPE, brand_name = "mujosh",
  start_date = "2020-10-10", end_date = "2020-10-31", category = c("frame", "sunglasses")
)

}
