% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sales-data.R
\encoding{UTF-8}
\name{get_sales_data}
\alias{get_sales_data}
\title{get sales data from BI database}
\usage{
get_sales_data(con, brand_name, start_date, end_date, ...)
}
\arguments{
\item{con}{BI con connector}

\item{brand_name}{事业部名称}

\item{start_date}{销售数据开始日期}

\item{end_date}{销售数据截至日期}

\item{...}{汇总字段 默认为空 可指定任意门店属性或商品属性字段}
}
\value{
sales data
}
\description{
获取销售数据 有数量、金额、吊牌金额三个指标，维度需自行指定
}
\examples{
get_sales_data
get_sales_data(con,brand_name = 'MUJOSH',start_date = '2020-01-01',end_date = '2020-10-25',SHOP_NO,SKU_NO)

}
