% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stock-data.R
\encoding{UTF-8}
\name{get_stock_data}
\alias{get_stock_data}
\title{get stock data from BI database}
\usage{
get_stock_data(
  con,
  brand_name,
  stock_date,
  goods_categories = c("镜架", "太阳镜"),
  ...
)
}
\arguments{
\item{con}{BI con connector}

\item{brand_name}{事业部}

\item{stock_date}{库存日期 默认昨天}

\item{goods_categories}{分析大类}

\item{...}{需要添加汇总字段,默认年,月,SHOP_NO,SKU_NO}
}
\value{
data frame
}
\description{
从BI数仓中获取库存数据 按照年、月、SHOP_NO,SKU_NO汇总
}
\examples{
dt <- get_stock_data(con = con,brand_name = 'mujosh',stock_date = lubridate::today()-days(1),goods_categories = c('闀滄灦','澶槼闀<9c>'),鍒嗘瀽澶х被,浠撲綅绫诲瀷)


}
