% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\encoding{UTF-8}
\name{store}
\alias{store}
\title{ghzy}
\usage{
store(con)
}
\arguments{
\item{con}{BI con connector}
}
\value{
a dataframe
}
\description{
get the store information from BI database
}
\examples{
store <- store(con)
}
