% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\encoding{UTF-8}
\name{store}
\alias{store}
\title{store information}
\usage{
store(
  con,
  brand_name,
  channel_type = NULL,
  area_name = NULL,
  boss_name = NULL,
  shop_no = NULL
)
}
\arguments{
\item{con}{BI con connector}

\item{brand_name}{brand_name}

\item{channel_type}{store type}

\item{area_name}{the area of store}

\item{shop_no}{the shop of Number}

\item{boo_name}{the boss of store}
}
\value{
a dbplyrlink
}
\description{
get the store information from BI database
}
\details{
门店属性:SHOP_NO 门店名称 原ERP店编码 品牌 一级部门 门店性质 国家 管辖区域 省份 城市 城市等级 门店负责人 区域经理 经营状态
店铺类型 老板
}
\examples{
store <- store(con,brand_name = "mujosh")
store <- store(con,brand_name = c("mujosh",'aojo'),channel_type= 'zhiying')
store_table <- store(con,brand_name = "mujosh") \%>\% collect()
}
